Fix missing license in query and allow empty namespace
Root cause was returning mapped super classes from the ScanEntity Spring JPA
repository.

- Add check before starting queued scan to only scan if it was not scanned
before. (This fixes duplicate scans of already queued packages.)

+++ src/main/java/com/philips/research/licensescanner/core/domain/LicenseInteractor.java
@@ -66,18 +66,22 @@ public class LicenseInteractor implements LicenseService {
         try {
             LOG.info("Scan license for {}:{} {} from {}", origin, name, version, location);
             final var pkg = getOrCreatePackage(origin, name, version);
-            path = downloader.download(location);
-            //TODO Check hash after download
-            scan = store.createScan(pkg, location);
-            detector.scan(path, scan, licenseThreshold);
-            LOG.info("Detected license for {}:{} {} is '{}'", origin, name, version, scan.getLicense());
+            if (store.latestScan(pkg).isEmpty()) {
+                path = downloader.download(location);
+                //TODO Check hash after download
+                scan = store.createScan(pkg, location);
+                detector.scan(path, scan, licenseThreshold);
+                LOG.info("Detected license for {}:{} {} is '{}'", origin, name, version, scan.getLicense());
+            }
         } catch (Exception e) {
             LOG.error("Scanning failed", e);
             if (scan != null) {
                 scan.setError(e.getMessage());
             }
         } finally {
-            deleteDirectory(path);
+            if (path != null) {
+                deleteDirectory(path);
+            }
         }
     }
